feat: extract re-used type and utils  (option, item, item clicked)
