+++ src/components/messenger/lib/types.ts
@@ -1,0 +1,11 @@
+export interface Item {
+  id: string;
+  name: string;
+  image?: string;
+}
+
+export interface Option {
+  value: string;
+  label: string;
+  image?: string;
+}
