+++ src/components/messenger/list/start-group-panel.tsx
@@ -2,7 +2,9 @@ import * as React from 'react';
 
 import { Button } from '@zero-tech/zui/components';
 
-import { AutocompleteMembers, Option } from '../autocomplete-members';
+import { Option } from '../lib/types';
+
+import { AutocompleteMembers } from '../autocomplete-members';
 import { PanelHeader } from './panel-header';
 import { SelectedUserTag } from './selected-user-tag';
